requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
yt-dlp==2024.10.22
rq==1.16.2
redis==5.0.8
pydantic-settings==2.6.1
qdrant-client==1.11.3
python-dotenv==1.0.1
httpx==0.27.2

# Security & Authentication
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
starlette==0.40.0

# Monitoring & Logging
prometheus-fastapi-instrumentator==7.0.0
structlog==24.5.0
rich==13.9.4
psutil==6.1.1

# Database & Validation
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
email-validator==2.2.0

# Caching & Performance
aiocache==0.12.2
aioredis==2.0.1

# Optional (install later if needed)
# faster-whisper==1.0.3
# openai==1.52.2
